FetchWeatherData/requirements.txt
flask
geopy
requests
google-cloud-storage